+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 **Changed**
 
 * Changed provider settings to have it's own component instead of overwriting `heptacom-admin-open-auth-client-edit-page`
-* Changed `view/administration/index-js.html.twig` to only check for login form route matches
+* Changed `view/administration/index-js.html.twig` to only check for relevant route matches
+* Changed `view/administration/index-js.html.twig` to allow login in `sw-inactivity-login`
 
 # 4.2.1